I give you a working stocker.
I haven't made it build by default yet because theres a few issues that need looking into...

First, it doesn't show up in the blist right away, you need to unload/reload to make that happen.
Second, the format of the output sucks, we need to camel case the names, and figure out how to handle the rest, it's just weird you have to see it.
The parsing code is extremely fragile
The colors could use tweaking
Needs a good review for error/leak/spell checking